requirements.txt
tweepy>=4.15.0
python-dotenv>=1.1.0
supabase>=2.15.0
requests>=2.32.3
asyncio>=3.4.3
resend>=0.6.0
openai>=1.0.0 